라우트변경및 스타트 컴포넌트 분리
